requirements.txt
matplotlib
scikit-image
scikit-learn
imgaug
yacs
segmentation-models-pytorch
torchsummary
numpy
